Make TagMap a final class, and remove logging.
